Extract ConfigStorage
